Add config option for trailing closure rewriting
Default to the automatic expansion introduced in
90c124c369fa2938908508ee4c540033f95761e3 but allow the user to disable
it via a configuration file.

This is expressed as an enumeration rather than a boolean flag because
there is room for other levels of rewriting. E.g. for trailing closures,
a "basic" level might be implemented as rewriting to a pair of brackets
on the same line, with a single placeholder, rather than the "full"
multi-line behavior.