Fix a bug in `TaskScheduler` that incorrectly restricted the number of executing tasks
When you have execution limits of 2 high priority tasks and 1 low priority tasks, the intention was that you could either execute 2 high priority tasks or 1 low priority task. The way it was implemented right now, we didn’t take the priority of the currently executing tasks into account and wouldn’t schedule the low-priority task at all if there was a high-priority task already executing. Fix this and also clarify the doc comment on `maxConcurrentTasksByPriority`.