CMake: Fix SwiftCrypto/SwiftASN1 order
SwiftCrypto requires SwiftASN1. SwiftCrypto should be declaring the
dependency in its import files, but is not currently doing so resulting
in `find_package(SwiftCrypto)` failing because the `SwiftASN1` target
does not yet exist.
By placing `find_package(SwiftASN1)` before finding SwiftCrypto, we
ensure that the `SwiftASN1` target is populated ahead of time and
finding SwiftCrypto succeeds.