Compare against expected output for formatting tests
The formatting request tests currently test for the specific list
of changes returned by the server, but that list is tied to the
differences algorithm used by the standard library, which isn't
guaranteed to be stable. This change compares the expected result
with a new string generated by applying the list of changes, which
is stable.