Fix issue that caused background indexing to be disabled when `sourcekit-lsp` is launched without options
When launching sourcekit-lsp without any command-line arguments, we would set `backgroundIndexing = false` in the options. Unless the user overwrites this somehow, this means that background indexing is disabled.

This is not an issue in VS Code, because it explicitly enables background indexing in the initialization request but for all other editors this means that background indexing was likely disabled by default.

Simply remove that line since `backgroundIndexing` defaults to `true` by now anyway.